Fix large amount of size file cannot read
refs #7494

+++ Arduino15/packages/SPRESENSE/hardware/spresense/1.0.0/libraries/SDHCI/SDHCI.h
@@ -124,7 +124,7 @@ public:
   * @param [in] nbyte The number of elements in buf.
   * @return The total number of bytes successfully read, or -1 if none is available 
   */
-  int read(void *buf, uint16_t nbyte);
+  int read(void *buf, size_t nbyte);
 
  /**
   * @brief Seek to a new position in the file.
